PP/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.15
pymysql==1.0.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==0.27.2
